refactor: refactor user profile management with update and change password functionality
