+++ features/profile/api/use-update-user.ts
@@ -1,0 +1,33 @@
+import { axiosInstance } from "@/lib/axios";
+import { useMutation } from "@tanstack/react-query";
+
+export const useUpdateUser = ({
+  onSuccess,
+  onError,
+}: {
+  onSuccess?: () => void;
+  onError?: (e: Error) => void;
+}) => {
+  return useMutation({
+    mutationFn: async ({
+      data,
+      userId,
+    }: {
+      data: FormData;
+      userId: string;
+    }) => {
+      const { data: response } = await axiosInstance.patch(
+        `/users/${userId}`,
+        data,
+        {
+          headers: {
+            "Content-Type": "multipart/form-data",
+          },
+        },
+      );
+      return response;
+    },
+    onSuccess,
+    onError,
+  });
+};
